Use zlib compression and binary patches for cartridge rom images
Please note that for now, this feature is only implemented for the
u2plus as a proof of concept.

Using the minz and bsdiff libraries the cartridge rom images are now
stored in compressed form in flash memory. In addition, the ntsc
variants are each stored as binary patches to their respective pal
variants. This method reduces the amount of flash space required for
cartridge images by 288k, from 568k down to 280k.

The miniz and bsdiff sources have been added and two fields have been
added to the t_flash_address structure:

uint32_t uncompressed_size
uint8_t base_id

An uncompressed_size greater than 0 denotes a compressed image. If
base_id is greater than 0, this image is interpreted as a binary patch
based on the image denoted by base_id.

Decompression and patching is implemented in private methods in
w25q_flash.cc only. The read_image() function now checks the
additional fields and decompresses and patches the image contents if
appropriate.

In order to generate the compressed and patched images, some simple
command line tools have been added and integrated into the build
process. The new Makefile in roms/ defines the set of roms to be
included, generates the images and also creates the image address
definitions and the list of files to be included in w25q_flash.cc and
update_binaries.s respectively.

The existing t_flash_address definitions have been adjusted to include
the new fields to make sure they are properly initialized everywhere,
even though compressed/patched images are only used in u2plus for now.