Enabled z-machine for u2 as well, use sprintf instead of snprintf
snprintf is not available on microblaze for a standalone application
like mb_ultimate. Using it generates errors when linking.

Please note that on the u2, this branch currently results in an application
binary of 770k, which exceeds the maximum allowed flash size of 768k, so
don't update this version to flash. It can be run from the filesystem,
though.